# Code/Server/docs/swagger.yaml
basePath: /v1
definitions:
...
  models.ArchiveRequest:
    properties:
      archive:
        type: boolean
    type: object
...
  models.IdResponse:
    properties:
      id:
        type: string
    type: object
...
  models.InviteRequest:
    properties:
      end_date:
        type: string
      start_date:
        type: string
      tenant_email:
        type: string
    required:
    - start_date
    - tenant_email
    type: object
...
  models.PropertyInventoryResponse:
    properties:
      address:
        type: string
      apartment_number:
        type: string
      archived:
        type: boolean
      area_sqm:
        type: number
      city:
        type: string
      country:
        type: string
      created_at:
        type: string
      deposit_price:
        type: number
      id:
        type: string
      invite:
        $ref: '#/definitions/models.propertyInviteResponse'
      lease:
        $ref: '#/definitions/models.propertyLeaseResponse'
      name:
        type: string
      nb_damage:
        type: integer
      owner_id:
        type: string
      picture_url:
        type: string
      postal_code:
        type: string
      rental_price_per_month:
        type: number
      rooms:
        items:
          $ref: '#/definitions/models.roomResponse'
        type: array
      status:
        $ref: '#/definitions/models.PropertyStatus'
    type: object
  models.PropertyRequest:
    properties:
      address:
        type: string
      apartment_number:
        type: string
      area_sqm:
        type: number
      city:
        type: string
      country:
        type: string
      deposit_price:
        type: number
      name:
        type: string
      postal_code:
        type: string
      rental_price_per_month:
        type: number
    required:
    - address
    - area_sqm
    - city
    - country
    - deposit_price
    - name
    - postal_code
    - rental_price_per_month
    type: object
  models.PropertyResponse:
    properties:
      address:
        type: string
      apartment_number:
        type: string
      archived:
        type: boolean
      area_sqm:
        type: number
      city:
        type: string
      country:
        type: string
      created_at:
        type: string
      deposit_price:
        type: number
      id:
        type: string
      invite:
        $ref: '#/definitions/models.propertyInviteResponse'
      lease:
        $ref: '#/definitions/models.propertyLeaseResponse'
      name:
        type: string
      nb_damage:
        type: integer
      owner_id:
        type: string
      picture_url:
        type: string
      postal_code:
        type: string
      rental_price_per_month:
        type: number
      status:
        $ref: '#/definitions/models.PropertyStatus'
    type: object
  models.PropertyStatus:
    enum:
    - unavailable
    - invite sent
    - available
    type: string
    x-enum-varnames:
    - StatusUnavailable
    - StatusInviteSent
    - StatusAvailable
  models.PropertyUpdateRequest:
    properties:
      address:
        type: string
      apartment_number:
        type: string
      area_sqm:
        type: number
      city:
        type: string
      country:
        type: string
      deposit_price:
        type: number
      name:
        type: string
      postal_code:
        type: string
      rental_price_per_month:
        type: number
    type: object
...
  models.furnitureResponse:
    properties:
      archived:
        type: boolean
      id:
        type: string
      name:
        type: string
      quantity:
        type: integer
    type: object
  models.propertyInviteResponse:
    properties:
      end_date:
        type: string
      start_date:
        type: string
      tenant_email:
        type: string
    type: object
  models.propertyLeaseResponse:
    properties:
      active:
        type: boolean
      end_date:
        type: string
      id:
        type: string
      start_date:
        type: string
      tenant_email:
        type: string
      tenant_name:
        type: string
    type: object
  models.roomResponse:
    properties:
      archived:
        type: boolean
      furnitures:
        items:
          $ref: '#/definitions/models.furnitureResponse'
        type: array
      id:
        type: string
      name:
        type: string
    type: object
  utils.Error:
    properties:
      code:
        $ref: '#/definitions/utils.ErrorCode'
      error:
        type: string
    type: object
  utils.ErrorCode:
    enum:
    - invalid-password
    - cannot-fetch-user
    - user-not-found
    - cannot-create-user
    - no-claims
    - cannot-decode-user
    - missing-fields
    - missing-file
    - cannot-hash-password
    - email-already-exists
    - test-error
    - too-many-requests
    - invite-not-found
    - user-must-have-same-email-as-invite
    - invite-already-exists-for-email-or-property
    - lease-already-exists-for-tenant-and-property
    - property-not-found
    - property-is-not-yours
    - not-an-owner
    - not-a-tenant
    - user-already-exists-as-owner
    - property-already-exists
    - property-not-available
    - tenant-already-has-lease
    - no-active-lease
    - lease-not-found
    - cannot-end-non-current-lease
    - inventory-report-must-be-linked-to-current-lease
    - no-pending-lease
    - document-not-found
    - damage-not-found
    - damage-already-exists
    - cannot-update-fixed-damage
    - damage-already-fixed
    - failed-upload-file
    - failed-delete-file
    - bad-base64-string
    - property-picture-not-found
    - user-profile-picture-not-found
    - room-not-found
    - room-already-exists
    - room-not-in-this-property
    - furniture-not-found
    - furniture-already-exists
    - furniture-not-in-this-property
    - inventory-report-already-exists
    - cannot-modify-inventory-report
    - inventory-report-not-found
    - room-state-already-exists
    - furniture-state-already-exists
    - invalid-object-type
    - error-request-chatgpt-api
    - failed-send-email
    type: string
    x-enum-varnames:
    - InvalidPassword
    - CannotFetchUser
    - UserNotFound
    - CannotCreateUser
    - NoClaims
    - CannotDecodeUser
    - MissingFields
    - MissingFile
    - CannotHashPassword
    - EmailAlreadyExists
    - TestError
    - TooManyRequests
    - InviteNotFound
    - UserSameEmailAsInvite
    - InviteAlreadyExists
    - LeaseAlreadyExist
    - PropertyNotFound
    - PropertyNotYours
    - NotAnOwner
    - NotATenant
    - UserAlreadyExistsAsOwner
    - PropertyAlreadyExists
    - PropertyNotAvailable
    - TenantAlreadyHasLease
    - NoActiveLease
    - LeaseNotFound
    - CannotEndNonCurrentLease
    - InvReportMustBeCurrentLease
    - NoLeaseInvite
    - DocumentNotFound
    - DamageNotFound
    - DamageAlreadyExists
    - CannotUpdateFixedDamage
    - DamageAlreadyFixed
    - FailedUploadFile
    - FailedDeleteFile
    - BadBase64String
    - PropertyPictureNotFound
    - UserProfilePictureNotFound
    - RoomNotFound
    - RoomAlreadyExists
    - RoomNotInThisProperty
    - FurnitureNotFound
    - FurnitureAlreadyExists
    - FurnitureNotInThisProperty
    - InventoryReportAlreadyExists
    - CannotModifyInventoryReport
    - InventoryReportNotFound
    - RoomStateAlreadyExists
    - FurnitureStateAlreadyExists
    - InvalidObjectType
    - ErrorRequestChatGPTAPI
    - FailedSendEmail
host: localhost:3001
info:
  contact:
    email: martin.d-herouville@epitech.eu
    name: Mazettt
  description: This is the API used by the Immotep application.
  termsOfService: http://swagger.io/terms/
  title: Immotep API
  version: "1.0"
paths:
...
  /owner/properties/:
    get:
      consumes:
      - application/json
      description: Get properties information of an owner, optionally filtered by
        archive status
      parameters:
      - description: 'Filter by archive status (default: false)'
        in: query
        name: archive
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: List of properties
          schema:
            items:
              $ref: '#/definitions/models.PropertyResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Error'
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: Get properties of an owner
      tags:
      - property
    post:
      consumes:
      - application/json
      description: Create a new property for an owner
      parameters:
      - description: Property data
        in: body
        name: property
        required: true
        schema:
          $ref: '#/definitions/models.PropertyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created property ID
          schema:
            $ref: '#/definitions/models.IdResponse'
        "400":
          description: Missing fields
          schema:
            $ref: '#/definitions/utils.Error'
        "409":
          description: Property already exists
          schema:
            $ref: '#/definitions/utils.Error'
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: Create a new property
      tags:
      - property
  /owner/properties/{property_id}/:
    get:
      consumes:
      - application/json
      description: Get property information by its ID
      parameters:
      - description: Property ID
        in: path
        name: property_id
        required: true
        type: string
      - description: 'ONLY FOR OWNER: optional Lease ID (default: current)'
        in: query
        name: lease_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Property data
          schema:
            $ref: '#/definitions/models.PropertyResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Error'
        "403":
          description: Property not yours
          schema:
            $ref: '#/definitions/utils.Error'
        "404":
          description: Property not found
          schema:
            $ref: '#/definitions/utils.Error'
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: Get property by ID
      tags:
      - property
    put:
      consumes:
      - application/json
      description: Update property information by its ID
      parameters:
      - description: Property ID
        in: path
        name: property_id
        required: true
        type: string
      - description: Property data
        in: body
        name: property
        required: true
        schema:
          $ref: '#/definitions/models.PropertyUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated property ID
          schema:
            $ref: '#/definitions/models.IdResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Error'
        "403":
          description: Property not yours
          schema:
            $ref: '#/definitions/utils.Error'
        "404":
          description: Property not found
          schema:
            $ref: '#/definitions/utils.Error'
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: Update property by ID
      tags:
      - property
  /owner/properties/{property_id}/archive:
    put:
      consumes:
      - application/json
      description: Toggle archive status of a property by its ID
      parameters:
      - description: Property ID
        in: path
        name: property_id
        required: true
        type: string
      - description: Archive status
        in: body
        name: archive
        required: true
        schema:
          $ref: '#/definitions/models.ArchiveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Property archive status
          schema:
            $ref: '#/definitions/models.IdResponse'
        "400":
          description: Mising fields
          schema:
            $ref: '#/definitions/utils.Error'
        "403":
          description: Property not yours
          schema:
            $ref: '#/definitions/utils.Error'
        "404":
          description: Property not found
          schema:
            $ref: '#/definitions/utils.Error'
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: Toggle archive property by ID
      tags:
      - property
  /owner/properties/{property_id}/cancel-invite:
    delete:
      consumes:
      - application/json
      description: Cancel pending lease invite
      parameters:
      - description: Property ID
        in: path
        name: property_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Invite canceled
        "403":
          description: Property is not yours
          schema:
            $ref: '#/definitions/utils.Error'
        "404":
          description: No pending lease
          schema:
            $ref: '#/definitions/utils.Error'
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: Cancel invite
      tags:
      - property
...
  /owner/properties/{property_id}/inventory/:
    get:
      consumes:
      - application/json
      description: Get property information by its ID with inventory including rooms
        and furnitures
      parameters:
      - description: Property ID
        in: path
        name: property_id
        required: true
        type: string
      - description: 'ONLY FOR OWNER: optional Lease ID (default: current)'
        in: query
        name: lease_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Property inventory data
          schema:
            $ref: '#/definitions/models.PropertyInventoryResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Error'
        "403":
          description: Property not yours
          schema:
            $ref: '#/definitions/utils.Error'
        "404":
          description: Property not found
          schema:
            $ref: '#/definitions/utils.Error'
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: Get property inventory by ID
      tags:
      - property
      - inventory
...
  /owner/properties/{property_id}/picture/:
    put:
      consumes:
      - multipart/form-data
      description: Update property's picture
      parameters:
      - description: Property ID
        in: path
        name: property_id
        required: true
        type: string
      - description: Property picture
        in: formData
        name: picture
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Updated property ID
          schema:
            $ref: '#/definitions/models.IdResponse'
        "400":
          description: Missing fields or bad base64 string
          schema:
            $ref: '#/definitions/utils.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Error'
        "403":
          description: Property not yours
          schema:
            $ref: '#/definitions/utils.Error'
        "404":
          description: Property not found
          schema:
            $ref: '#/definitions/utils.Error'
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: Update property's picture
      tags:
      - property
...
  /owner/properties/{property_id}/send-invite:
    post:
      consumes:
      - application/json
      description: Invite tenant to owner's property
      parameters:
      - description: Property ID
        in: path
        name: property_id
        required: true
        type: string
      - description: Invite params
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.InviteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created invite ID
          schema:
            $ref: '#/definitions/models.IdResponse'
        "400":
          description: Missing fields
          schema:
            $ref: '#/definitions/utils.Error'
        "403":
          description: Property is not yours
          schema:
            $ref: '#/definitions/utils.Error'
        "404":
          description: Property not found
          schema:
            $ref: '#/definitions/utils.Error'
        "409":
          description: Invite already exists for this email
          schema:
            $ref: '#/definitions/utils.Error'
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: Invite tenant to owner's property
      tags:
      - property
...
  /tenant/leases/{lease_id}/property/:
    get:
      consumes:
      - application/json
      description: Get property information by its ID
      parameters:
      - description: Lease ID or `current`
        in: path
        name: lease_id
        required: true
        type: string
      - description: 'ONLY FOR OWNER: optional Lease ID (default: current)'
        in: query
        name: lease_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Property data
          schema:
            $ref: '#/definitions/models.PropertyResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Error'
        "403":
          description: Property not yours
          schema:
            $ref: '#/definitions/utils.Error'
        "404":
          description: Property not found
          schema:
            $ref: '#/definitions/utils.Error'
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: Get property by ID
      tags:
      - property
  /tenant/leases/{lease_id}/property/inventory/:
    get:
      consumes:
      - application/json
      description: Get property information by its ID with inventory including rooms
        and furnitures
      parameters:
      - description: Lease ID or `current`
        in: path
        name: lease_id
        required: true
        type: string
      - description: 'ONLY FOR OWNER: optional Lease ID (default: current)'
        in: query
        name: lease_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Property inventory data
          schema:
            $ref: '#/definitions/models.PropertyInventoryResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Error'
        "403":
          description: Property not yours
          schema:
            $ref: '#/definitions/utils.Error'
        "404":
          description: Property not found
          schema:
            $ref: '#/definitions/utils.Error'
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: Get property inventory by ID
      tags:
      - property
      - inventory
...
securityDefinitions:
  Bearer:
    description: Enter the token with the `Bearer ` prefix, e.g. "Bearer abcde12345".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"